% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_cbind}
\alias{rcpp_cbind}
\title{rcpp_cbind}
\usage{
rcpp_cbind(valores2, matriz)
}
\arguments{
\item{A}{será el vector elegido como ejemplo.}

\item{B}{la primera matriz.}

\item{matriz}{Matriz de ejemplo para la función.}

\item{valores2}{Es el vector que se unirá a la matriz anterior.}
}
\value{
Una nueva matriz con los datos del vector y la primera matriz separados por columnas.
}
\description{
Función que une una matriz y un vector en una nueva matriz cuando
el tamaño del vector sea igual a la cantidad de filas de la primera matriz.
}
\details{
[[Rcpp::export]]
}
\examples{
matriz <- matrix(10:24, 3, 5)
matriz
valores2 <- c(2, 54, 32)
rcpp_cbind(valores2, matriz)
}
