% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_maxmin}
\alias{rcpp_maxmin}
\title{rcpp_maxmin}
\usage{
rcpp_maxmin(matriz, matriz2)
}
\arguments{
\item{A}{será la primera matriz elegida como ejemplo.}

\item{B}{la segunda matriz.}
}
\value{
Una nueva matriz con la comparación realizada.
}
\description{
Función que compara dos matrices desde la primera matriz con su primera
fila a la primera columna de la segunda matriz y en vez de multiplicar se elige el
numero más pequeño y luego va tomando el más alto.
}
\details{
[[Rcpp::export]]
}
\examples{
# Actividad 1
matriz = matrix(5:20, nrow=4, ncol=4)
matriz2 = matrix(10:25, nrow=4, ncol=4)
# Operador no lineal convolución maxmin
rcpp_maxmin(matriz, matriz2)
}
