% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_multimatrix}
\alias{rcpp_multimatrix}
\title{rcpp_multimatrix}
\usage{
rcpp_multimatrix(matriz, matriz2)
}
\arguments{
\item{A}{ser치 la primera matriz elegida como ejemplo.}

\item{B}{la segunda matriz.}
}
\value{
Una nueva matriz con la multiplicaci칩n realizada.
}
\description{
Funci칩n que multiplica matrices dependiendo de que la cantidad de
columnas de la primera matriz sea igual a la cantidad de filas de la segunda matriz.
}
\details{
[[Rcpp::export]]
}
\examples{
# Multiplicaci칩n de Matrices
matriz = matrix(5:14, nrow=2, ncol=5)
matriz2 = matrix(10:29, nrow=5, ncol=4)
# Multiplicar Matrices
rcpp_multimatrix(matriz, matriz2)
}
