% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_rbind}
\alias{rcpp_rbind}
\title{rcpp_rbind}
\usage{
rcpp_rbind(valores, matriz)
}
\arguments{
\item{A}{será el vector elegido como ejemplo.}

\item{B}{la primera matriz.}

\item{matriz}{Matriz de ejemplo para la función.}

\item{valores}{Es el vector que se unirá a la matriz anterior.}
}
\value{
Una nueva matriz con los datos del vector y la primera matriz separados por filas.
}
\description{
Función que une una matriz y un vector en una nueva matriz cuando
el tamaño del vector sea igual a la cantidad de columnas de la primera matriz.
}
\details{
[[Rcpp::export]]
}
\examples{
# cbind y rbind
matriz <- matrix(10:24, 3, 5)
matriz
valores <- matrix(c(2, 54, 32, 65, 12), 1, 5)
rcpp_rbind(valores, matriz)
}
