% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_whichmax}
\alias{rcpp_whichmax}
\title{rcpp_whichmax}
\usage{
rcpp_whichmax(matriz)
}
\arguments{
\item{A}{será la matriz elegida como ejemplo.}
}
\value{
Devolucion de posicion donde se ubica el numero más grande dentro de la matriz.
}
\description{
Función que recorre una matriz buscando la ubicación donde se encuentra
el número más grande de esta.
}
\details{
[[Rcpp::export]]
}
\examples{
# Función which.min y which.max
valores <- c(40, 56, 87, 78, 2, 1, 54, 45, 87, 34, 79, 65)
matriz <- matrix(valores, nrow = 3, ncol = 4)
rcpp_whichmax(matriz)
}
