% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rcpp_whichmin}
\alias{rcpp_whichmin}
\title{rcpp_whichmin}
\usage{
rcpp_whichmin(matriz)
}
\arguments{
\item{A}{será la matriz elegida como ejemplo.}

\item{matriz}{Matriz de ejemplo para la función.}

\item{valores}{Vector de datos que se utilizará para crear la matriz.}
}
\value{
Devolucion de posicion donde se ubica el numero más pequeño dentro de la matriz.
}
\description{
Función que recorre una matriz buscando la ubicación donde se encuentra
el número más pequeño de esta.
}
\details{
[[Rcpp::export]]
}
\examples{
# Función which.min
valores <- c(40, 56, 87, 78, 2, 1, 54, 45, 87, 34, 79, 65)
matriz <- matrix(valores, nrow = 3, ncol = 4)
rcpp_whichmin(matriz)
}
